JobApplicationTracker/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.27
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
email-validator==2.1.0
pytest==7.4.0 